Split shutdown into sched_shutdown() for later and do_shutdown() for immediate. Now uses multiple runq's for better scheduling semantics:
The scheduler is implemented as a FIFO, two-level priority, circular
queue with the following priorities:

\begin{enumerate}
\item{Consume Queue (high priority): Data objects to be processed}
\item{Produce Queue (low priority): Modules that can produce new data objects}
\end{enumerate}

Each entry in the Consume Queue is a data object and the
module that needs to process it.  Rather than scheduling by time
quanta, we schedule by data object.  Once a module has completed
processing a data object, it voluntarily returns to the scheduler.  As
a result, there are no attempts to explicitly enforce fairness or
priorities.

The Produce Queue is used to keep track of modules that can produce
data.  Some modules, typically at the roots of graphs, produce data.
If a graph is expected to produce its own data, then the scheduler is
told to seed the Produce Queue with the root(s) of the graph.  Each
time a module produces data, it signals through its return code
whether it can produce more data.  If it can, then the module is
placed back on the tail of the Produce Queue.  Elements are taken from
the Produce Queue only when the Consume queue is empty.  Thus the
system is biased towards processing existing data before acquiring new
data.

When a module requests to be shutdown, it should not receive any
additional data objects or additional produce calls.  Modules are also
shutdown when all of their children are shutdown (which is reminiscent
of a Unix SIGPIPE condition).  In these cases, the module is
immediately shutdown.  However, a third case occurs when all of the
parents of a module are shutdown.  In this case a shutdown action (an
entry with no data object) is enqueued at the end of the Consume
Queue.  The module will be shutdown only when this entry is dequeued,
and that is guaranteed to occur only after the module has processed
any other objects that were previously enqueued for it before the
parents were shutdown.